Issue #308: Rename RawFile
