+++ src/south/FolderScanner/FolderScanner.class.js
@@ -5,11 +5,11 @@ const ProtocolHandler = require('../ProtocolHandler.class')
 const databaseService = require('../../services/database.service')
 
 /**
- * Class RawFile
+ * Class FolderScanner
  */
-class RawFile extends ProtocolHandler {
+class FolderScanner extends ProtocolHandler {
   /**
-   * Constructor for RawFile
+   * Constructor for FolderScanner
    * @constructor
    * @param {Object} dataSource - The data source
    * @param {Engine} engine - The engine
@@ -30,7 +30,7 @@ class RawFile extends ProtocolHandler {
     if (this.preserveFiles) {
       const { engineConfig } = this.engine.configService.getConfig()
       const databasePath = `${engineConfig.caching.cacheFolder}/${this.dataSource.dataSourceId}.db`
-      this.database = await databaseService.createRawFilesDatabase(databasePath)
+      this.database = await databaseService.createFolderScannerDatabase(databasePath)
     }
   }
 
@@ -99,7 +99,7 @@ class RawFile extends ProtocolHandler {
 
     await Promise.all(filenames.map(async (filename) => {
       const stats = fs.statSync(path.join(this.inputFolder, filename))
-      const modified = await databaseService.getRawFileModifyTime(this.database, filename)
+      const modified = await databaseService.getFolderScannerModifyTime(this.database, filename)
       if (modified) {
         if ((stats.mtimeMs > modified)) {
           filesToHandle.push(filename)
@@ -139,10 +139,10 @@ class RawFile extends ProtocolHandler {
 
     this.logger.debug(`Upsert handled file ${filename} with modify time ${stats.mtimeMs}`)
 
-    await databaseService.upsertRawFile(this.database, filename, stats.mtimeMs)
+    await databaseService.upsertFolderScanner(this.database, filename, stats.mtimeMs)
   }
 }
 
-RawFile.schema = require('./schema')
+FolderScanner.schema = require('./schema')
 
-module.exports = RawFile
+module.exports = FolderScanner
